Moves text substitutor logic into hope pkg.
