+++ cmd/hope/remove.go
@@ -62,7 +62,7 @@ var removeCmd = &cobra.Command{
 			switch resourceType {
 			case hope.ResourceTypeFile:
 				if len(resource.Parameters) != 0 {
-					content, err := utils.ReplaceParametersInFile(resource.File, resource.Parameters)
+					content, err := hope.ReplaceParametersInFile(resource.File, resource.Parameters)
 					if err != nil {
 						return err
 					}
@@ -84,7 +84,7 @@ var removeCmd = &cobra.Command{
 				log.Trace(inline)
 
 				if len(resource.Parameters) != 0 {
-					inline, err = utils.ReplaceParametersInString(inline, resource.Parameters)
+					inline, err = hope.ReplaceParametersInString(inline, resource.Parameters)
 					if err != nil {
 						return err
 					}
